@@ -8,16 +8,21 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
+        "ext-gd": "*",
         "algolia/algoliasearch-client-php": "^3.3",
         "algolia/scout-extended": "^3.0",
         "aws/aws-sdk-php": "^3.304",
+        "bepsvpt/blurhash": "^3.0",
         "datomatic/nova-enum-field": "^1.6",
         "digital-creative/nova-slider-filter": "^1.0",
         "eventsauce/eventsauce": "^3.5",
+        "eventsauce/message-repository-for-illuminate": "^1.2",
         "guzzlehttp/guzzle": "^7.2",
+        "hammerstone/sidecar": "^0.4.2",
         "inertiajs/inertia-laravel": "^0.6.3",
         "inspheric/nova-indicator-field": "^1.43",
+        "jenssegers/imagehash": "^0.9.0",
         "kra8/laravel-snowflake": "^2.2",
         "laravel/framework": "^10.10",
         "laravel/horizon": "^5.20",
@@ -39,7 +44,7 @@
         "raiym/instagram-php-scraper": "^0.12.9",
         "spatie/eloquent-sortable": "^4.0",
         "spatie/enum": "^3.13",
-        "spatie/laravel-data": "^3.6",
+        "spatie/laravel-data": "^4",
         "spatie/laravel-event-sourcing": "^7.3",
         "spatie/laravel-ignition": "^2.4",
         "spatie/laravel-model-states": "^2.4",
@@ -49,8 +54,11 @@
         "spatie/laravel-tags": "^4.4",
         "spatie/laravel-typescript-transformer": "^2.3",
         "spatie/laravel-view-models": "^1.5",
+        "spatie/mjml-sidecar": "^1.0",
         "spatie/nova-tags-field": "^4.0",
         "spatie/typescript-transformer": "^2.2",
+        "symfony/http-client": "^7.1",
+        "symfony/postmark-mailer": "^7.1",
         "tightenco/ziggy": "^1.0",
         "vyuldashev/nova-money-field": "^2.0"
     },
@@ -60,7 +68,7 @@
         "eventsauce/test-utilities": "^3.4",
         "fakerphp/faker": "^1.9.1",
         "laravel/breeze": "^1.21",
-        "laravel/dusk": "^7.11",
+        "laravel/dusk": "^8.2",
         "laravel/pint": "^1.0",
         "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
@@ -74,7 +82,14 @@
     "autoload": {
         "psr-4": {
             "App\\": "src/App/",
-            "Domain\\": "src/Domains/",
+            "Domain\\Jobs\\": "src/Domains/Jobs",
+            "Domain\\Leads\\": "src/Domains/Leads",
+            "Domain\\Present\\": "src/Domains/Present",
+            "Domain\\Profiles\\": "src/Domains/Profiles",
+            "Domain\\Work\\": "src/Domains/Work",
+            "Domain\\Work2\\": "src/Domains/Work2/src",
+            "Tests\\Work2\\": "src/Domains/Work2/tests",
+
             "Support\\": "src/Support/",
             "Database\\Factories\\": "database/factories/",
             "Database\\Seeders\\": "database/seeders/"
@@ -108,6 +123,7 @@
         "aws/aws-sdk-php": [
             "Ses",
             "S3",
+            "Iam",
             "Sqs",
             "Lambda",
             "Ec2",
