@@ -2,33 +2,72 @@
 
 namespace App\Http\Controllers;
 
-use Domain\Jobs\Data\ApplicationData;
+use App\Mail\CleanMail;
+use App\Mail\ClientSubmittedPreference;
+use Domain\Jobs\Data\ListingData;
+use Domain\Present\Data\PresentationData;
 use Domain\Present\Models\Presentation;
-use Domain\Work\Models\Application;
+use Domain\Work2\Models\Listing;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Mail;
 use Inertia\Inertia;
 
 class PresentationController extends Controller
 {
     public function show(Presentation $presentation)
     {
-        $role = $presentation->role;
-        $role->load("job");
-        $applications = Application::whereIn("id", $presentation->applications)->orderBy('order_column')->get();
-        $applications->load("model.digitals", "casting_photos", "photos", "casting_videos", "model.portfolio");
+        $presentation->load("role.job");
 
-        return Inertia::render('Roles/Proposal')
-            ->with("role", $role)
-            ->with("presentation", $presentation)
-            ->with("applications", ApplicationData::collection($applications));
+        $listings = $presentation->presentationListings()
+                ->orderBy('order_column')
+                ->with("listing.model.digitals", "listing.casting_videos", "listing.model.portfolio")
+                ->get()
+                ->pluck("listing");
+
+        return Inertia::render('Roles/Presentation')
+            ->with("presentation", PresentationData::from($presentation))
+            ->with("listings", ListingData::collect($listings));
     }
 
-    public function shortlist(Presentation $presentation)
+    public function favorite(Presentation $presentation, Request $request)
     {
-        foreach (Application::whereIn('id', request()->get("applications"))->get() as $application) {
-            $application->shortlisted_at = $application->shortlisted_at ? null : now();
-            $application->save();
+        $favorites = $request->get("favoriteListings");
+
+        $allListings = $presentation->presentationListings()
+            ->orderBy('order_column')
+            ->with("listing.model.digitals", "listing.casting_videos", "listing.model.portfolio")
+            ->get()
+            ->pluck("listing");
+
+        $favoritedListings = $allListings->filter(function ($listing) use ($favorites) {
+            return in_array($listing->id, $favorites);
+        });
+
+        $names = $favoritedListings->pluck('model.name');
+
+        foreach ($allListings as $listing) {
+            if ($listing->favorited_at && !in_array($listing->id, $favorites)) {
+                $listing->favorited_at = null;
+            }
+
+            if (!$listing->favorited_at && in_array($listing->id, $favorites)) {
+                $listing->favorited_at = now();
+            }
+
+            $listing->save();
         }
 
-        return "Ok";
+        Mail::to($presentation->role->job->responsible_user)
+            ->send(new CleanMail(
+                messageSubject: "Client has submitted favorites",
+                messageContent: [
+                    "The client has submitted preferences for the following models:",
+                    ...$names
+                ],
+                actionText: "View presentation",
+                actionUrl: route("presentations.show", $presentation)
+            ));
+
+        return Inertia::render('Roles/PrelistSubmitted')->with("role", $presentation->role);
     }
 }
