@@ -2,37 +2,38 @@
 
 namespace App\Notifications;
 
-use App\Notifications\SidemailData\SidemailNotification;
-use App\Notifications\SidemailData\SidemailRecipient;
+use App\Mail\CleanMail;
 use Domain\Profiles\Models\Model;
-use Domain\Work\Models\Application;
+use Domain\Work2\Models\Listing;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Mail\Mailable;
 use Illuminate\Notifications\Notification;
 
-class Shortlisted extends Notification implements ShouldQueue, SidemailNotification
+class Shortlisted extends Notification implements ShouldQueue
 {
     use Queueable;
 
-    public function __construct(public Application $application) {}
+    public function __construct(public Listing $listing)
+    {
+    }
 
     public function via(object $notifiable): string
     {
-        return SidemailChannel::class;
+        return 'mail';
     }
 
-    public function toSideMail(Model|Authenticatable $notifiable): SideMailMessage
+    public function toMail(Model|Authenticatable $notifiable): Mailable
     {
-        return new SideMailMessage(
-            recipient: SidemailRecipient::fromModel($notifiable),
-            template: 'shortlisted',
-            data: [
-                'role' => $this->application->role->name,
-                'job' => $this->application->role->job->title,
-                'brand' => $this->application->role->job->brand->name,
-                'url' => route('applications.show', $this->application),
-            ]
-        );
+        return (new CleanMail(
+            'You have been shortlisted',
+            [
+                "Hi {$this->listing->model->first_name},",
+                "You have been shortlisted for the role {$this->listing->role->name} for {$this->listing->role->job->title}.",
+            ],
+            'View Role',
+            route('roles.show', $this->listing->role),
+        ))->to($notifiable->email);
     }
 }
