@@ -2,8 +2,8 @@
 
 namespace App\Nova\Actions;
 
-use Domain\Jobs\Models\Invite;
 use Domain\Jobs\Models\Role;
+use Domain\Work2\Actions\Invite;
 use Illuminate\Bus\Queueable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Support\Collection;
@@ -26,12 +26,13 @@ class InviteForRole extends Action
     public function handle(ActionFields $fields, Collection $models)
     {
         foreach ($models as $model) {
-            $invite = new Invite();
-            $invite->role_id = $fields->role_id;
-            $invite->model_id = $model->id;
-            $invite->save();
 
-            Action::message("Invite for role has been sent to " . $models->first()->first_name . " " . $models->first()->last_name);
+            app(Invite::class)->execute(
+                Role::find($fields->role_id),
+                $model
+            );
+
+            Action::message("Invite for role has been sent to " . $model->name);
         }
     }
 
