@@ -3,11 +3,11 @@
 namespace App\Nova\Actions;
 
 use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Support\Collection;
 use Laravel\Nova\Actions\Action;
 use Laravel\Nova\Fields\ActionFields;
+use Laravel\Nova\Fields\Text;
 use Laravel\Nova\Fields\Textarea;
 use Laravel\Nova\Http\Requests\NovaRequest;
 
@@ -24,8 +24,13 @@ class Reject extends Action
      */
     public function handle(ActionFields $fields, Collection $collection)
     {
-        foreach ($collection as $application) {
-            app(\Domain\Work\Actions\Reject::class)->execute($application, $fields->message);
+        foreach ($collection as $listing) {
+            app(\Domain\Work2\Actions\Reject::class)
+                ->execute(
+                    $listing,
+                    $fields->subject,
+                    $fields->message
+                );
         }
     }
 
@@ -38,6 +43,7 @@ class Reject extends Action
     public function fields(NovaRequest $request)
     {
         return [
+            Text::make('Subject')->default("Sorry, you've not been selected"),
             Textarea::make('Message', 'message')
                 ->help("Use no salutation, include the signature.")
                 ->rules('required'),
