@@ -13,10 +13,13 @@ use Laravel\Nova\Fields\Text;
 use Laravel\Nova\Fields\Textarea;
 use Laravel\Nova\Http\Requests\NovaRequest;
 use Laravel\Nova\Panel;
+use Outl1ne\NovaSortable\Traits\HasSortableManyToManyRows;
 use Vyuldashev\NovaMoneyField\Money;
 
 class Role extends Resource
 {
+    use HasSortableManyToManyRows;
+
     public static $model = \Domain\Jobs\Models\Role::class;
 
     public function title()
@@ -70,16 +73,15 @@ class Role extends Resource
                 ]
             ),
             Text::make("Travel reimbursement note")->hideFromIndex(),
-            Text::make('Invites', function() {
+            HasMany::make("Listings"),
+            HasMany::make("Presentations"),
+            Text::make('Listings', function() {
                 return '<div style="display: flex; width: 400px; height: 120px; overflow-x: scroll; overflow-y: hidden">
-                    ' .implode("", $this->invites->map(function ($invite) {
-                        return '<img src="'.$invite->model->profile_picture_cdn.'?w=720&h=960&fit=crop&fm=auto&crop=faces" title="'.$invite->model->name.'" width="90" height="120" />';
+                    ' .implode("", $this->listings->map(function ($listing) {
+                        return '<img src="'.$listing->model->profile_picture_cdn.'?w=720&h=960&fit=crop&fm=auto&crop=faces" title="'.$listing->model->name.'" width="90" height="120" />';
                     })->toArray())
                     . '</div>';
             })->asHtml()->onlyOnIndex(),
-            HasMany::make("Applications"),
-            HasMany::make("Invites", "open_invites"),
-            HasMany::make("Presentations"),
             Text::make('Public URL', function() {
                 return '<a href="'.route("roles.show", $this->id).'" target="_blank">'.route("roles.show", $this->id).'</a>';
             })->asHtml()->onlyOnDetail(),
