@@ -2,7 +2,6 @@
 
 namespace Domain\Profiles\Data;
 
-use Spatie\LaravelData\Attributes\Validation\Rule;
 use Spatie\LaravelData\Data;
 
 /** @typescript */
@@ -12,6 +11,7 @@ class ModelSocialsData extends Data
         public ?string $instagram,
         public ?string $tiktok,
         public ?string $website,
+        public ?string $showreel_link,
     ) {
     }
 }
