@@ -12,7 +12,8 @@ class PhotoData extends Data
 {
     public function __construct(
         public int $id,
-        public string $path
+        public string $path,
+        public ?string $hash
     )
     { }
 }
