@@ -2,9 +2,6 @@
 
 namespace Domain\Profiles\Data;
 
-use App\Transformers\CdnPathTransformer;
-use Spatie\LaravelData\Attributes\Computed;
-use Spatie\LaravelData\Attributes\WithTransformer;
 use Spatie\LaravelData\Data;
 
 /** @typescript */
@@ -13,7 +10,7 @@ class VideoData extends Data
     public function __construct(
         public int $id,
         public string $path,
-        public string $mux_id,
+        public ?string $mux_id,
     )
     { }
 }
