@@ -6,12 +6,53 @@ use Domain\Profiles\Actions\AnalysePhoto;
 use Domain\Profiles\Models\Model;
 use Domain\Profiles\Models\Photo;
 use Illuminate\Contracts\Auth\Authenticatable;
-use Illuminate\Support\Facades\Storage;
 use Support\Actions\DeletePhoto;
 use Support\Actions\MovePhoto;
+use Support\Actions\PhashPhoto;
 
 class PhotoRepository
 {
+    public function update(\Illuminate\Database\Eloquent\Model $model, string $folder, $photos)
+    {
+        $newSort = collect($photos)->map(function($photo) use($folder, $model) {
+
+            if (array_key_exists("deleted", $photo) && $photo['deleted']===true) {
+
+                if ($photoObj = Photo::find($photo['id'])) {
+                    app(DeletePhoto::class)->onQueue()->execute($photoObj->path);
+                    $photoObj->delete();
+                }
+
+                return null;
+            }
+
+            if (array_key_exists("isNew", $photo) && $photo['isNew']===true) {
+
+                $photoObj = new Photo;
+                $photoObj->photoable()->associate($model);
+                $photoObj->path = str_replace("tmp/", "photos/", $photo['path']);
+                $photoObj->folder = $folder;
+                $photoObj->save();
+
+                if ($photoObj->wasRecentlyCreated) {
+                    app(MovePhoto::class)->onQueue()->execute($photo['path'], $photoObj->path);
+
+                    if ($folder === Photo::FOLDER_ACTIVITIES || $folder === Photo::FOLDER_WORK_EXPERIENCE) {
+                        app(AnalysePhoto::class)->onQueue()->execute($photoObj);
+                    }
+                    app(PhashPhoto::class)->onQueue()->execute($photoObj);
+                }
+                return $photoObj->id;
+
+            }
+
+            return $photo['id'];
+        });
+
+
+        Photo::setNewOrder($newSort->toArray());
+    }
+
     /**
      * @param Model|Authenticatable $model
      */
@@ -26,46 +67,5 @@ class PhotoRepository
                     "deleteRoute" => route('account.photos.delete', $photo),
                 ];
             });
-    }
-
-    public function update(\Illuminate\Database\Eloquent\Model|Authenticatable $model, string $folder, $photos)
-    {
-        $newSort = collect($photos)->map(function($photo) use($folder, $model) {
-
-            if (!isset($photo['isNew'])) {
-                if (isset($photo['deleted']) && $photo['deleted'] != 0) {
-
-                    if ($photoObj = Photo::find($photo['id'])) {
-                        app(DeletePhoto::class)->onQueue()->execute($photoObj->path);
-                        $photoObj->delete();
-                    }
-
-                    return null;
-                }
-
-                return $photo['id'];
-            }
-
-            if (isset($photo['deleted']) && $photo['deleted']!=0) return null;
-
-            $photoObj = new Photo;
-            $photoObj->photoable()->associate($model);
-            $photoObj->path = str_replace("tmp/", "photos/", $photo['path']);
-            $photoObj->folder = $folder;
-            $photoObj->save();
-
-            if ($photo['path'] != $photoObj->path) {
-                app(MovePhoto::class)->onQueue()->execute($photo['path'], $photoObj->path);
-
-                if ($folder === Photo::FOLDER_ACTIVITIES || $folder === Photo::FOLDER_WORK_EXPERIENCE) {
-                    app(AnalysePhoto::class)->onQueue()->execute($photoObj);
-                }
-            }
-
-            return $photoObj->id;
-        });
-
-
-        Photo::setNewOrder($newSort->toArray());
     }
 }
