@@ -2,8 +2,9 @@
 
 namespace Support;
 
-use Illuminate\Foundation\Auth\User as Authenticatable;
+use Database\Factories\UserFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Laravel\Sanctum\HasApiTokens;
 
@@ -19,7 +20,17 @@ class User extends Authenticatable
         'password',
     ];
 
-    public function isSuperUser()
+    public static function newFactory()
+    {
+        return UserFactory::new();
+    }
+
+    public function canImpersonate()
+    {
+        return true;
+    }
+
+    public function isSuperAdmin()
     {
         return str_starts_with($this->email, 'bas@') || str_starts_with($this->email, 'jeroen@');
     }
