@@ -4,7 +4,6 @@ namespace Tests\Unit\Domains\Profiles\Models;
 
 use Domain\Profiles\Models\Model;
 use Domain\Profiles\Models\Photo;
-use Illuminate\Support\Facades\DB;
 use Tests\TestCase;
 
 class ModelTest extends TestCase
@@ -13,10 +12,10 @@ class ModelTest extends TestCase
     {
         $model = Model::factory()->create();
         $model->photos()->createMany([
-            ['analysis' => ['hair' => ['blond', 'red']], 'path' => 'test', 'folder' => Photo::FOLDER_ACTIVITIES],
-            ['analysis' => ['hair' => ['blond', 'black']], 'path' => 'test', 'folder' => Photo::FOLDER_ACTIVITIES],
+            ['analysis' => ['hair' => 'blond, red'], 'path' => 'test', 'folder' => Photo::FOLDER_ACTIVITIES],
+            ['analysis' => ['hair' => 'blond, black'], 'path' => 'test', 'folder' => Photo::FOLDER_ACTIVITIES],
         ]);
 
-        $this->assertEquals(['hair' => ['blond', 'red', 'black']], $model->toSearchableArray()['photo_analysis']);
+        $this->assertEquals(['hair' => 'blond, red, black'], $model->toSearchableArray()['photo_values']);
     }
 }
